Moved HologramEvent into events package
